scripts/generators/docker-compose/requirements.txt
deepmerge==2.0
inflect==7.5.0
Jinja2==3.1.4
PyYAML==6.0.2
pytest==8.3.4